fix : 리뷰쓰기 수정 #40

+++ app/src/main/java/com/example/fit_i/data/model/request/ReviewRequest.kt
@@ -1,7 +1,7 @@
 package com.example.fit_i.data.model.request
 
-data class ReviewRequest (
-    val trainerIdx : Int,
-    val grade : Int,
-    val contents : String
+data class ReviewRequest(
+    val trainerIdx: Long,
+    val grade: Int,
+    val contents: String
 )
